fix: the component prop doesn't refresh when the state is updated
